Demo 3, Step 4: Mock defined by Copilot Edits
I used the following prompt (and retained the code generated, without any intervention):

Create a mock codeunit for this interface and put it into the "test" folder. I need the mock to have stubs for all methods that return values, and spies for all other methods. I want to be able to spy on if a method is invoked, and on all of its parameters. I want stubbing methods to be prefixed "SetResult_" and spy methods to be prefixed with "IsInvoked_" and I want all private variables to be prefixed with "_".